chore: revamping folder structure (#264)
