+++ drivers/mysql/internal/backfill.go
@@ -7,11 +7,11 @@ import (
 	"sort"
 	"time"
 
-	"github.com/datazip-inc/olake/logger"
 	"github.com/datazip-inc/olake/pkg/jdbc"
 	"github.com/datazip-inc/olake/protocol"
 	"github.com/datazip-inc/olake/types"
 	"github.com/datazip-inc/olake/utils"
+	"github.com/datazip-inc/olake/utils/logger"
 )
 
 // backfill implements full refresh sync mode for MySQL
