+++ pkg/waljs/types.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/datazip-inc/olake/protocol"
 	"github.com/datazip-inc/olake/types"
-	"github.com/datazip-inc/olake/typeutils"
+	"github.com/datazip-inc/olake/utils/typeutils"
 	"github.com/jackc/pglogrepl"
 )
