+++ protocol/check.go
@@ -3,9 +3,9 @@ package protocol
 import (
 	"fmt"
 
-	"github.com/datazip-inc/olake/logger"
 	"github.com/datazip-inc/olake/types"
 	"github.com/datazip-inc/olake/utils"
+	"github.com/datazip-inc/olake/utils/logger"
 	"github.com/spf13/cobra"
 )
