+++ protocol/sync.go
@@ -5,9 +5,9 @@ import (
 	"strings"
 	"sync"
 
-	"github.com/datazip-inc/olake/logger"
 	"github.com/datazip-inc/olake/types"
 	"github.com/datazip-inc/olake/utils"
+	"github.com/datazip-inc/olake/utils/logger"
 	"github.com/spf13/cobra"
 )
