+++ utils/jsonschema/generator/schema_attrs.go
@@ -8,8 +8,8 @@ import (
 	"unicode"
 	"unicode/utf8"
 
-	"github.com/datazip-inc/olake/jsonschema/schema"
-	"github.com/datazip-inc/olake/logger"
+	"github.com/datazip-inc/olake/utils/jsonschema/schema"
+	"github.com/datazip-inc/olake/utils/logger"
 )
 
 func (g *JSONSchemaGenerator) shouldReturnRef(decl *declInfo) bool {
