+++ utils/jsonschema/jsonschema.go
@@ -6,8 +6,8 @@ import (
 
 	"github.com/goccy/go-json"
 
-	"github.com/datazip-inc/olake/jsonschema/generator"
-	"github.com/datazip-inc/olake/jsonschema/schema"
+	"github.com/datazip-inc/olake/utils/jsonschema/generator"
+	"github.com/datazip-inc/olake/utils/jsonschema/schema"
 	"sigs.k8s.io/yaml"
 )
