+++ utils/jsonschema/schema/schema_numeric.go
