+++ utils/jsonschema/schema/schema_simple.go
