+++ utils/typeutils/flatten.go
