+++ writers/iceberg/config.go
@@ -5,8 +5,8 @@ import (
 	"os"
 	"strings"
 
-	"github.com/datazip-inc/olake/logger"
 	"github.com/datazip-inc/olake/utils"
+	"github.com/datazip-inc/olake/utils/logger"
 )
 
 // CatalogType represents supported Iceberg catalog implementations
