+++ writers/iceberg/iceberg_utils.go
@@ -14,8 +14,8 @@ import (
 	"time"
 
 	"github.com/datazip-inc/olake/constants"
-	"github.com/datazip-inc/olake/logger"
 	"github.com/datazip-inc/olake/utils"
+	"github.com/datazip-inc/olake/utils/logger"
 	"github.com/datazip-inc/olake/writers/iceberg/proto"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/credentials/insecure"
